# backend/app/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GP Appointment API",
    "version": "1.0.0",
    "description": "API to manage appointments with doctors"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://api.gpbooking.icu"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateScheduleDTO": {
        "type": "object",
        "properties": {
          "DoctorId": {
            "type": "string",
            "description": "The ID of the doctor"
          },
          "DoctorName": {
            "type": "string",
            "description": "The name of the doctor"
          },
          "Date": {
            "type": "string",
            "format": "date",
            "description": "The date of the schedule"
          },
          "StartTime": {
            "type": "string",
            "format": "time",
            "description": "The start time of the schedule"
          },
          "EndTime": {
            "type": "string",
            "format": "time",
            "description": "The end time of the schedule"
          }
        }
      },
...
      "CreateAppointmentDTO": {
        "type": "object",
        "properties": {
          "ScheduleId": {
            "type": "string",
            "description": "The ID of the schedule"
          },
          "FirstName": {
            "type": "string",
            "description": "The first name of the patient"
          },
          "LastName": {
            "type": "string",
            "description": "The last name of the patient"
          },
          "DateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "The date of birth of the patient"
          },
          "Phone": {
            "type": "string",
            "description": "The phone number of the patient"
          },
          "Email": {
            "type": "string",
            "description": "The email of the patient"
          },
          "Address": {
            "type": "string",
            "description": "The address of the patient"
          }
        }
      },
      "AppointmentProperties": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateAppointmentDTO"
          },
          {
            "$ref": "#/components/schemas/CreateScheduleDTO"
          },
          {
            "type": "object",
            "properties": {
              "BookingReference": {
                "type": "string",
                "description": "The booking reference of the appointment"
              },
              "Status": {
                "type": "string",
                "description": "The status of the appointment",
                "enum": [
                  "active",
                  "cancelled",
                  "completed"
                ]
              }
            }
          }
        ]
      },
      "AppointmentDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AppointmentProperties"
          },
          {
            "type": "object",
            "properties": {
              "Id": {
                "type": "string",
                "description": "The ID of the appointment"
              },
              "Version": {
                "type": "integer",
                "description": "The version of the record"
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
    "/admin/appointments": {
      "get": {
        "summary": "Get a list of appointments",
        "description": "Get a list of appointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "doctorId",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The doctor ID"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The start date of the appointment"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The end date of the appointment"
          }
        ],
        "responses": {
          "200": {
            "description": "return a list of appointments based on the query"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admin/appointments/{appointmentId}": {
      "delete": {
        "summary": "Delete an appointment by ID",
        "description": "Delete an appointment by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointmentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the appointment to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Appointment deleted successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admin/appointments/{appointmentId}/complete": {
      "post": {
        "summary": "Complete an appointment",
        "description": "Complete an appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointmentId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The appointment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "return the completed appointment"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/public/appointments": {
      "get": {
        "summary": "Get an appointment by booking reference, last name, and date of birth",
        "description": "Get an appointment by booking reference, last name, and date of birth",
        "parameters": [
          {
            "in": "query",
            "name": "reference",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The booking reference"
          },
          {
            "in": "query",
            "name": "lastname",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The last name of the patient"
          },
          {
            "in": "query",
            "name": "dob",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The date of birth of the patient"
          }
        ],
        "responses": {
          "200": {
            "description": "return a list of appointments based on the query"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "description": "Create a new appointment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDTO",
                "required": [
                  "ScheduleId",
                  "LastName",
                  "DateOfBirth",
                  "Email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return the created appointment"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/user/appointments": {
      "get": {
        "summary": "List appointments",
        "description": "List appointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "doctorId",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The doctor ID"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The start date of the appointment"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "The end date of the appointment"
          }
        ],
        "responses": {
          "200": {
            "description": "return a list of appointments based on the query"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "description": "Create a new appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDTO",
                "required": [
                  "ScheduleId",
                  "LastName",
                  "DateOfBirth",
                  "Email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return the created appointment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/user/appointments/{appointmentId}": {
      "get": {
        "summary": "Get an appointment by ID",
        "description": "Get an appointment by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointmentId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The appointment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "return the appointment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDTO"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update an appointment",
        "description": "Update an appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointmentId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The appointment ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "return the updated appointment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDTO"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}